prefixed and extrapolated classes into separate files and some backwards compatibility, updated tests
Separated logger classes into multiple files to keep with common OOP PHP library development.

Prefixed classes with Loggr so as to not pollute global space with common terms such as Event or Events, etc.

Added static functions for semi-backwards compatibility. In test.php an example showed LogClient() created Loggr::LogClient, but of course Loggr_LogClient could be used as well.
