+++ loggr-php/loggr_data_type.php
@@ -1,0 +1,7 @@
+<?php
+
+class Loggr_DataType
+{
+    const html = 0;
+    const plaintext = 1;
+}
