+++ loggr-php/loggr_event.php
@@ -1,0 +1,29 @@
+<?php
+
+class Loggr_Event
+{
+	public $text;
+	public $source;
+	public $user;
+	public $link;
+	public $data;
+	public $value;
+	public $tags;
+	public $latitude;
+	public $longitude;
+	public $dataType = Loggr_DataType::plaintext;
+
+	public function __get($attribute) {
+		$attribute = lcfirst($attribute);
+		if(isset($this->{$attribute})) {
+			return $this->{$attribute};
+		}
+	}
+
+	public function __set($attribute, $value) {
+		$attribute = lcfirst($attribute);
+		if(isset($this->{$attribute})) {
+			$this->{$attribute} = $value;
+		}
+	}
+}
